test/mcp-demo/requirements.txt
fastmcp
torch
torchvision
Pillow
opencv-python
numpy
scikit-learn
matplotlib
seaborn
flask
flask-cors
werkzeug
requests
tqdm
volcengine
easyocr
openai
python-dotenv

